+++ CHANGES.md
@@ -2,5 +2,6 @@
 
 ## Version 0.0.1
 
+-- Initial version